w3c-reports/12092014/gtk/runned-tests/websockets/constructor/011-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/handshake_protocol' failed: Error during WebSocket handshake: Sec-WebSocket-Protocol mismatch

FAIL WebSockets: protocol mismatch assert_true: got open expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/websockets/constructor/011.html:19:16
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
